Reconfigured parts of the project. Rewritten NN code.
Renamed various classes and files to make it more clear what they are.
Completely rewritten the NN code so that it now does two things:
a) It trains two different networks, one player player 1 and one for player 2. This solves a vew issues to do with asymetric games.
b) Has a variable number of hidden layers.